+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## 1.1.0
+* `GameClient#advanceUntil()` を追加
+
 ## 1.0.1
 * `GameContextParameterObject#verbose` を追加